Moved the game search/adding to list to a separate component
Moved the game search/adding to list to a separate component since the gameList component was getting very long with all of that in it. So now the gameList just has the list and button to delete the list.
